move constant to settings
