+++ WCF-Generator/App.config
@@ -19,9 +19,26 @@
   <!-- Cofiguration of generation wcf client-->
   <wcfClientGenerator>
     <wcfServices>
-      <add ClientInterfaceName = "IBackOffice" ServiceInterfaceFileName = "IBackOfficeService.cs" />
-      <add ClientInterfaceName = "ITerminal" ServiceInterfaceFileName = "IService.cs"/>
-      <add ClientInterfaceName = "IOnlineService" ServiceInterfaceFileName = "IOnlineService.cs"/>
+      <add ClientInterfaceName = "IBackOffice" 
+           ServiceInterfaceFileName = "IBackOfficeService.cs" 
+           TargetProjectName ="YumaPos.Client.WCF"
+           FaultNamespace ="YumaPos.Shared.API.Faults"
+           ApiInterfaceProjectName="YumaPos.Shared.Infrastructure"
+           ApiInterfaceProjectFolder = "API"/>
+      
+      <add ClientInterfaceName = "ITerminal"
+           ServiceInterfaceFileName = "IService.cs" 
+           TargetProjectName ="YumaPos.Client.WCF"
+           FaultNamespace = "YumaPos.Shared.API.Faults"
+           ApiInterfaceProjectName="YumaPos.Shared.Infrastructure"
+           ApiInterfaceProjectFolder = "API"/>
+      
+      <add ClientInterfaceName = "IOnlineService" 
+           ServiceInterfaceFileName = "IOnlineService.cs"
+           TargetProjectName ="YumaPos.Client.WCF"
+           FaultNamespace ="YumaPos.Shared.API.Faults"
+           ApiInterfaceProjectName="YumaPos.Shared.Infrastructure"
+           ApiInterfaceProjectFolder = "API"/>
     </wcfServices>
   </wcfClientGenerator>
   <!-- Cofiguration of generation repositories-->
